Break up helpers.tsx and reorganize new code
