--- frontend/src/components/AlertModal.tsx
@@ -1,43 +1,0 @@
-import { FunctionComponent, useEffect } from "react";
-import Button from "react-bootstrap/Button";
-import Modal from "react-bootstrap/Modal";
-
-export const AlertModal: FunctionComponent<{
-  show: boolean;
-  title: string;
-  content: string | null;
-  onHide: () => void;
-}> = ({ show, title, content, onHide }) => {
-  useEffect(() => {
-    if (!show) return;
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key === "Enter") {
-        e.preventDefault();
-        onHide();
-      }
-    };
-    window.addEventListener("keydown", handleKeyDown);
-    return () => {
-      window.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [show, onHide]);
-
-  return (
-    <Modal
-      show={show}
-      onHide={onHide}
-      aria-labelledby="contained-modal-title-vcenter"
-      centered
-    >
-      <Modal.Header closeButton>
-        <Modal.Title id="contained-modal-title-vcenter">{title}</Modal.Title>
-      </Modal.Header>
-      <Modal.Body>
-        <p>{content}</p>
-      </Modal.Body>
-      <Modal.Footer>
-        <Button onClick={onHide}>Ok</Button>
-      </Modal.Footer>
-    </Modal>
-  );
-};
