+++ frontend/src/components/DataGrid.tsx
@@ -1,13 +1,48 @@
 import { AgGridReact } from "ag-grid-react";
 import { AgGridReact as AgGridReactType } from "ag-grid-react/lib/agGridReact";
 import { RefObject, ClipboardEvent } from "react";
-import { CellEditRequestEvent, ColDef } from "ag-grid-community";
 import {
-  CACHE_BLOCK_SIZE,
-  defaultColDef,
-  SampleChange,
-} from "../shared/helpers";
+  CellEditRequestEvent,
+  ColDef,
+  ITooltipParams,
+} from "ag-grid-community";
 import { useNavigate } from "react-router-dom";
+import { createCustomHeader, lockIcon } from "../utils/gridFormatters";
+import { SampleChange } from "../types";
+import { CACHE_BLOCK_SIZE } from "../config";
+import { allEditableFields } from "../pages/samples/config";
+
+function getTooltipValue(params: ITooltipParams) {
+  if (!params.colDef || !("field" in params.colDef)) return undefined;
+  const field = params.colDef.field;
+  if (
+    (field === "cancerType" || field === "cancerTypeDetailed") &&
+    params.value === "N/A"
+  ) {
+    return (
+      "This code might have changed between different versions of the Oncotree API. " +
+      "For more details, visit oncotree.mskcc.org/mapping"
+    );
+  }
+  if (
+    allEditableFields.has(field!) &&
+    params.data?.sampleCategory === "clinical"
+  ) {
+    return "Clinical samples are not editable";
+  }
+  if (!allEditableFields.has(field!)) {
+    return "This column is read-only";
+  }
+}
+
+const defaultColDef: ColDef = {
+  sortable: true,
+  resizable: true,
+  editable: false,
+  headerComponentParams: createCustomHeader(lockIcon),
+  valueFormatter: (params) => (params.value === "null" ? "" : params.value),
+  tooltipValueGetter: (params: ITooltipParams) => getTooltipValue(params),
+};
 
 interface EditableGridProps {
   changes: Array<SampleChange>;
@@ -27,7 +62,7 @@ type DataGridPropsBase = {
   handleGridColumnsChanged: () => void;
 };
 
-export type DataGridProps = DataGridPropsBase &
+type DataGridProps = DataGridPropsBase &
   // Ensure that either all editing props are used or none are
   (EditableGridProps | NonEditableGridProps);
 
@@ -56,7 +91,6 @@ export function DataGrid({
           getChanges: () => changes,
           navigateFunction: navigate,
         }}
-        // TODO: put this obj somewhere more appropriate
         rowClassRules={{
           unlocked: (params) => params.data?.revisable === true,
           locked: (params) => params.data?.revisable === false,
