+++ frontend/src/components/DownloadButton.tsx
@@ -1,36 +1,7 @@
-import { ColDef } from "ag-grid-community";
 import { Button, ButtonGroup, Dropdown } from "react-bootstrap";
 import { CustomTooltip } from "./CustomToolTip";
 import InfoIcon from "@material-ui/icons/InfoOutlined";
-import { ReactNode } from "react";
-
-export interface DownloadOption {
-  /**
-   * Label for the download option button.
-   */
-  label: string;
-  /**
-   * Column definitions for the data to be downloaded.
-   * Typically, this will be the same column definitions used in the current table.
-   */
-  columnDefs: Array<ColDef>;
-  /**
-   * Async function that fetches the data to be downloaded.
-   * Usage: pass in `getRenderedData` from the `useDownload` hook to fetch the
-   * current search results on the page, or create your own function to fetch
-   * data for other purposes.
-   */
-  dataGetter: () => Promise<Array<any>>;
-  /**
-   * Optional content of a tooltip to display next to the download label
-   * in the dropdown menu.
-   */
-  tooltipContent?: string;
-  /**
-   * Optionally disable a download option in the dropdown menu.
-   */
-  disabled?: boolean;
-}
+import { DownloadOption } from "../hooks/useDownload";
 
 interface DownloadButtonProps {
   downloadOptions: Array<DownloadOption>;
@@ -45,7 +16,7 @@ export function DownloadButton({
     <Dropdown as={ButtonGroup}>
       {/* Main download button */}
       <Button size={"sm"} onClick={() => handleDownload(downloadOptions[0])}>
-        {downloadOptions[0].label}
+        {downloadOptions[0].buttonLabel}
       </Button>
 
       {downloadOptions.length > 1 && (
@@ -55,7 +26,7 @@ export function DownloadButton({
             {/* Dropdown download options */}
             {downloadOptions.map((downloadOption) => (
               <div
-                key={downloadOption.label}
+                key={downloadOption.buttonLabel}
                 className="d-flex align-items-center"
               >
                 <Dropdown.Item
@@ -63,13 +34,15 @@ export function DownloadButton({
                   onClick={() => handleDownload(downloadOption)}
                   disabled={downloadOption.disabled}
                 >
-                  {downloadOption.label}
+                  {downloadOption.buttonLabel}
                 </Dropdown.Item>
 
                 {/* Optional tooltip */}
-                <DropdownItemTooltip>
-                  {downloadOption.tooltipContent}
-                </DropdownItemTooltip>
+                {downloadOption.tooltipContent && (
+                  <DropdownItemTooltip>
+                    {downloadOption.tooltipContent}
+                  </DropdownItemTooltip>
+                )}
               </div>
             ))}
           </Dropdown.Menu>
@@ -79,8 +52,7 @@ export function DownloadButton({
   );
 }
 
-function DropdownItemTooltip({ children }: { children: ReactNode }) {
-  if (!children) return null;
+function DropdownItemTooltip({ children }: { children: string }) {
   return (
     <CustomTooltip
       icon={
