+++ frontend/src/components/FilterButtons.tsx
@@ -1,45 +1,32 @@
 import { Button, ButtonGroup } from "react-bootstrap";
 import { ColDef } from "ag-grid-community";
 import InfoIcon from "@material-ui/icons/InfoOutlined";
-import { CustomTooltip } from "../shared/components/CustomToolTip";
-import { ReactNode, useState } from "react";
+import { CustomTooltip } from "../components/CustomToolTip";
+import { useState } from "react";
 import { DashboardRecordContext } from "../generated/graphql";
 
 export type FilterButtonOption = {
   label: string;
-  columnDefs: ColDef[];
-  contexts?: DashboardRecordContext[];
+  columnDefs: Array<ColDef>;
+  contexts?: Array<DashboardRecordContext>;
 };
 
 interface FilterButtonsProps {
-  /**
-   * A map of filter button options, where the key is the button label and the
-   * value contains column definitions and optional contexts.
-   */
-  buttonOptions: Array<FilterButtonOption>;
-  /**
-   * Callback function to run additional logic when a filter button is clicked,
-   * like updating the column definitions.
-   */
-  onButtonClick: (filterButtonLabel: string) => void;
-  /**
-   * Content to be displayed inside the tooltip.
-   */
-  children: ReactNode;
+  options: Array<FilterButtonOption>;
+  onClick: (filterButtonLabel: string) => void;
+  children: string;
 }
 
 export function FilterButtons({
-  buttonOptions,
-  onButtonClick,
+  options,
+  onClick,
   children,
 }: FilterButtonsProps) {
-  const [activeButtonLabel, setActiveButtonLabel] = useState<string>(
-    buttonOptions[0].label
-  );
+  const [activeButtonLabel, setActiveButtonLabel] = useState(options[0].label);
 
   function handleButtonClick(filterButtonLabel: string) {
     setActiveButtonLabel(filterButtonLabel);
-    onButtonClick(filterButtonLabel);
+    onClick(filterButtonLabel);
   }
 
   return (
@@ -50,7 +37,7 @@ export function FilterButtons({
         {children}
       </CustomTooltip>{" "}
       <ButtonGroup>
-        {buttonOptions.map((buttonOption) => (
+        {options.map((buttonOption) => (
           <Button
             key={buttonOption.label}
             onClick={() => handleButtonClick(buttonOption.label)}
