+++ frontend/src/components/SearchBar.tsx
