+++ frontend/src/components/Toolbar.tsx
@@ -1,8 +1,7 @@
 import { ReactNode } from "react";
 import { Row } from "react-bootstrap";
 
-// TODO: change this back to Toolbar.tsx
-export function Toolbarr({ children }: { children: ReactNode }) {
+export function Toolbar({ children }: { children: ReactNode }) {
   return (
     <Row className="d-flex align-items-center justify-content-center my-2">
       {children}
