+++ frontend/src/hooks/useFetchData.ts
@@ -1,66 +1,44 @@
 import { RefObject, useCallback, useMemo, useState } from "react";
 import {
   AgGridSortDirection,
+  DashboardRecordColumnFilter,
   DashboardRecordSort,
+  DashboardRequestsDocument,
   DashboardSamplesQueryVariables,
 } from "../generated/graphql";
-import { useHookLazyGeneric } from "../shared/types";
 import { parseUserSearchVal } from "../utils/parseSearchQueries";
-import { getColumnFilters } from "../shared/helpers";
 import {
   IServerSideDatasource,
   IServerSideGetRowsParams,
 } from "ag-grid-community";
 import { AgGridReact as AgGridReactType } from "ag-grid-react/lib/agGridReact";
 import { usePhiEnabled } from "../contexts/PhiEnabledContext";
+import { LazyQueryHookOptions, useLazyQuery } from "@apollo/client";
+import { CACHE_BLOCK_SIZE } from "../config";
 
-export const CACHE_BLOCK_SIZE = 500; // number of rows to fetch at a time
+function useHookLazyGeneric(baseOptions?: LazyQueryHookOptions<any, any>) {
+  const options = { ...({} as const), ...baseOptions };
+  return useLazyQuery<any, any>(DashboardRequestsDocument, options);
+}
 
 interface UseFetchDataProps {
-  /**
-   * The lazy query hook generated by graphql-codegen.
-   */
   useRecordsLazyQuery: typeof useHookLazyGeneric;
-  /**
-   * The fieldName in the column definitions to use for the initial sort.
-   */
-  initialSortFieldName: string;
-  /**
-   * Additional contexts to filter records by that are always included in the query.
-   * For example, in the Request Samples view, the `igoRequestId` context is always included
-   * in each query to filter samples by the request they belong to.
-   */
-  contexts?: DashboardSamplesQueryVariables["contexts"];
-  /**
-   * Polling interval in milliseconds.
-   * https://www.apollographql.com/docs/react/data/queries#polling
-   */
-  pollInterval?: number;
-  /**
-   * The name of the query to use for fetching data matching that in the GraphQL schema.
-   * Examples: "dashboardRequests", "dashboardSamples", "dashboardPatients"
-   */
   queryName: string;
-  /**
-   * A reference to the AgGridReact component.
-   * Used to set the server-side datasource for the grid.
-   */
+  initialSortFieldName: string;
   gridRef: RefObject<AgGridReactType>;
-  /**
-   * The current user search value from the search box.
-   * Used to re-fetch data when the user performs a new search.
-   */
   userSearchVal: string;
+  contexts?: DashboardSamplesQueryVariables["contexts"];
+  pollInterval?: number;
 }
 
 export function useFetchData({
   useRecordsLazyQuery,
-  initialSortFieldName,
-  contexts = [],
-  pollInterval = 0, // 0 means no polling
   queryName,
+  initialSortFieldName,
   gridRef,
   userSearchVal,
+  contexts = [],
+  pollInterval = 0, // 0 means no polling
 }: UseFetchDataProps) {
   // Manage our own loading state becase the lazy query's provided `loading` state
   // does not toggle to `true` as `setServerSideDatasource` is running
@@ -152,4 +130,20 @@ export function useFetchData({
     startPolling,
     stopPolling,
   };
+}
+
+function getColumnFilters(
+  params: IServerSideGetRowsParams
+): Array<DashboardRecordColumnFilter> | undefined {
+  const filterModel = params.request.filterModel;
+  if (!filterModel || Object.keys(filterModel).length === 0) {
+    // All filter values are selected
+    return undefined;
+  }
+  return Object.entries(filterModel).map(([field, value]) => ({
+    field,
+    // Flexibly handle AG Grid's `any` type for filter settings by JSON.parse() this string value,
+    // then check the field name before consuming it at the GraphQL server (see https://stackoverflow.com/a/45601881)
+    filter: JSON.stringify(value),
+  }));
 }
