+++ frontend/src/types.ts
@@ -1,0 +1,14 @@
+import { ColDef, RowNode } from "ag-grid-community";
+
+export interface SampleChange {
+  primaryId: string;
+  fieldName: string;
+  oldValue: string;
+  newValue: string;
+  rowNode: RowNode;
+}
+
+export interface BuildDownloadOptionsParamsBase {
+  getCurrentData: () => Promise<Array<any>>;
+  currentColumnDefs: Array<ColDef<any>>;
+}
