+++ frontend/src/utils/dateFormatters.ts
@@ -1,0 +1,5 @@
+import moment from "moment";
+
+export function formatDate(date: moment.MomentInput) {
+  return date ? moment(date).format("YYYY-MM-DD") : null;
+}
